refactor(test): extract shared verbatim-warn spy helper
withWarnSpy + verbatimWarn were duplicated across three plan-mode test files
(plan-interactive, plan-refine, debate-strategy), crossing the project's
3-files-then-extract threshold (test-helpers.md). Move both to
test/helpers/verbatim-warn-spy.ts and import from the @test/helpers barrel.
